Fix local release builds for Android
The local release builds for Android in Fusion do not produce the same result as {N} CLI. The problem is in the `nativescript-dev-android-snapshot` plugin, which relies on a specific object, that's populated only when using CLI from the command line.
The plugin is fixed in it's latest version, so use it in the template. This way the builds will work in the same way in Fusion.

Fixes http://teampulse.telerik.com/view#item/330327